requirements.txt
# LangChain core
langchain
langchain-core
langchain-community
langchain-openai
langchain-pinecone
langchain-experimental

# Vector database
pinecone[grpc]
pinecone-client 

# Embeddings
sentence-transformers
huggingface-hub

# Document loading + splitting
pypdf

# OpenAI API
openai

Flask
python-dotenv

# Evaluation
ragas==0.2.10
datasets==3.2.0

# -e .